Bundle: split the vault settings dialog out of the account menu
