+++ app/apps/desktop/src/components/VaultSettingsDialog.tsx
@@ -1,0 +1,2903 @@
+/* The vault-settings dialog and its nine tabs — the whole settings surface of
+   the app, split out of `AccountMenu.tsx` so it can load on demand. Nothing
+   here is on the first screen: the sidebar footer (identity bar + popovers)
+   stays eager, and this chunk lands when someone actually opens settings. */
+import { useEffect, useMemo, useState } from "react";
+import {
+  type McpToolInfo,
+  type McpTokenRow,
+  type Member,
+  type MyBillingVault,
+  type OrgBilling,
+  type VaultCheckpoint,
+} from "../lib/api";
+import { toast } from "../lib/toast";
+import { agoFromIso, checkpointTitle, noteCountLabel } from "./versionFormat";
+import { ITEM_COLORS, itemColorValue } from "../lib/appearance";
+import { authManager } from "../lib/auth/authManager";
+import {
+  classifyLimitError,
+  type LimitKind,
+  limitFromError,
+  planPillLabel,
+  subscriptionStatusLine,
+  transferTargets,
+} from "../lib/billing";
+import * as ipc from "../lib/ipc";
+import {
+  checkForUpdate,
+  currentVersion,
+  installUpdate,
+  useUpdateState,
+} from "../lib/updater";
+import { readOrgVaults, useStore, type InviteResult } from "../store";
+import { buildInviteLink } from "../lib/inviteLink";
+import { SyncBadge } from "./Identity";
+import { AccessPanel } from "./AccessPanel";
+import { AsyncButton } from "./AsyncButton";
+import { ConfirmDialog } from "./ConfirmDialog";
+import { canActOnMember } from "./memberRoles";
+import { RoleSelect } from "./RoleSelect";
+// Static, not via ./Face: this module is itself a lazy chunk, so it pays for
+// the avatar chunk it is already loading.
+import { Avatar } from "./Avatar";
+import { MenuIcon } from "./MenuIcon";
+import { SettingsModal } from "./SettingsModal";
+import { Switch } from "./Switch";
+import { ThemeToggle } from "./ThemeToggle";
+import { formatPrice, perLabel, UpgradeDialog } from "./UpgradeDialog";
+import { useKnownOrgIds, useLocalVaults } from "./useVaultLists";
+
+export type SettingsTab =
+  | "general"
+  | "vaults"
+  | "members"
+  | "billing"
+  | "access"
+  | "mcp"
+  | "versioning"
+  | "import-export"
+  | "appearance"
+  | "updates";
+
+// Sections that only make sense once the vault is synced to an org. On a
+// local vault they're shown but locked, with a "Turn on sync" gate.
+//
+// Billing is deliberately NOT one of them (#109). Someone working in a local
+// folder can still own vaults that are billing, and a subscription left behind
+// by a DELETED vault has to be reachable from somewhere or the money is
+// unrecoverable. Only the per-vault card at the top of that tab needs a synced
+// vault, and it says so itself.
+const TEAM_TABS = new Set<SettingsTab>([
+  "members",
+  "access",
+  "mcp",
+  "versioning",
+]);
+
+/** General tab: name, folder, and sync state (incl. the Turn-on-sync CTA). */
+const GENERAL_TAB: { id: SettingsTab; label: string; icon: React.ReactNode } = {
+  id: "general",
+  label: "General",
+  icon: (
+    <MenuIcon>
+      <circle cx="12" cy="12" r="3" />
+      <path d="M12 2v2M12 20v2M4.9 4.9l1.4 1.4M17.7 17.7l1.4 1.4M2 12h2M20 12h2M4.9 19.1l1.4-1.4M17.7 6.3l1.4-1.4" />
+    </MenuIcon>
+  ),
+};
+
+const SETTINGS_TABS: Array<{ id: SettingsTab; label: string; icon: React.ReactNode }> = [
+  {
+    id: "vaults",
+    label: "Vaults",
+    icon: (
+      <MenuIcon>
+        <rect x="3" y="3" width="7" height="7" rx="1.5" />
+        <rect x="14" y="3" width="7" height="7" rx="1.5" />
+        <rect x="3" y="14" width="7" height="7" rx="1.5" />
+        <rect x="14" y="14" width="7" height="7" rx="1.5" />
+      </MenuIcon>
+    ),
+  },
+  {
+    id: "members",
+    label: "Members",
+    icon: (
+      <MenuIcon>
+        <path d="M17 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2" />
+        <circle cx="9" cy="7" r="4" />
+        <path d="M23 21v-2a4 4 0 0 0-3-3.87M16 3.13a4 4 0 0 1 0 7.75" />
+      </MenuIcon>
+    ),
+  },
+  {
+    id: "access",
+    label: "Access",
+    icon: (
+      <MenuIcon>
+        <rect x="4" y="11" width="16" height="10" rx="2" />
+        <path d="M8 11V7a4 4 0 0 1 8 0v4" />
+      </MenuIcon>
+    ),
+  },
+  {
+    id: "mcp",
+    label: "MCP",
+    icon: (
+      <MenuIcon>
+        <path d="M4 17l6-6-6-6" />
+        <path d="M12 19h8" />
+      </MenuIcon>
+    ),
+  },
+  {
+    id: "versioning",
+    label: "Versioning",
+    icon: (
+      <MenuIcon>
+        <path d="M3 12a9 9 0 1 0 2.6-6.4" />
+        <path d="M3 4v4h4" />
+        <path d="M12 8v4l3 2" />
+      </MenuIcon>
+    ),
+  },
+  {
+    id: "import-export",
+    label: "Import / Export",
+    icon: (
+      <MenuIcon>
+        <path d="M12 3v10" />
+        <path d="m8 9 4 4 4-4" />
+        <path d="M4 17v2a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2" />
+      </MenuIcon>
+    ),
+  },
+  {
+    id: "appearance",
+    label: "Appearance",
+    icon: (
+      <MenuIcon>
+        <circle cx="12" cy="12" r="10" />
+        <path d="M12 2a10 10 0 0 1 0 20 5 5 0 0 1 0-10 5 5 0 0 0 0-10" />
+      </MenuIcon>
+    ),
+  },
+  {
+    id: "updates",
+    label: "Updates",
+    icon: (
+      <MenuIcon>
+        <path d="M21 12a9 9 0 1 1-2.64-6.36" />
+        <path d="M21 3v5h-5" />
+      </MenuIcon>
+    ),
+  },
+];
+
+/** The Billing tab, inserted after Members only when the server has billing on. */
+const BILLING_TAB: { id: SettingsTab; label: string; icon: React.ReactNode } = {
+  id: "billing",
+  label: "Billing",
+  icon: (
+    <MenuIcon>
+      <rect x="2" y="5" width="20" height="14" rx="2" />
+      <path d="M2 10h20" />
+    </MenuIcon>
+  ),
+};
+
+/**
+ * Vault settings — a centered modal over the app (sharing its shell with
+ * Account settings via {@link SettingsModal}): everything about the vault lives
+ * here. Members (roster + join code + invites), Permissions (RBAC locks), and
+ * Appearance (theme + item colors).
+ */
+export function VaultSettingsDialog({
+  onClose,
+  onRequestSignIn,
+  initialTab,
+}: {
+  onClose: () => void;
+  onRequestSignIn?: () => void;
+  initialTab?: SettingsTab;
+}) {
+  const session = useStore((s) => s.session);
+  const organizations = useStore((s) => s.organizations);
+  const members = useStore((s) => s.members);
+  const billingConfig = useStore((s) => s.billingConfig);
+  const vault = useStore((s) => s.vault);
+  const syncEnabled = useStore((s) => s.syncEnabled);
+  const locals = useLocalVaults();
+
+  const [tab, setTab] = useState<SettingsTab>(initialTab ?? "general");
+
+  // Esc, click-away, focus and the backdrop all live in `SettingsModal`.
+  const activeOrg =
+    organizations.find((o) => o.id === session?.activeOrganizationId) ?? null;
+  // Is the vault we're looking at actually syncing to an org? A local
+  // vault (signed out, or an unsynced local folder) shows a reduced page
+  // with the team sections locked behind "Turn on sync".
+  const isSynced = syncEnabled && !!activeOrg;
+  const billingEnabled = billingConfig?.enabled === true;
+
+  // "Vaults" (switch/create/manage) is shown for an account OR when there
+  // are local folders to list — that's what "View all" opens into.
+  const showVaults = !!session || locals.length > 0;
+  const tabs = useMemo(() => {
+    const out = [GENERAL_TAB];
+    if (showVaults) out.push(...SETTINGS_TABS);
+    else out.push(...SETTINGS_TABS.filter((t) => t.id !== "vaults"));
+    if (billingEnabled) {
+      const idx = out.findIndex((t) => t.id === "members");
+      out.splice(idx >= 0 ? idx + 1 : out.length, 0, BILLING_TAB);
+    }
+    return out;
+  }, [showVaults, billingEnabled]);
+
+  if (!session && !vault) return null;
+  const myMember = members.find((m) => m.userId === session?.user.id);
+  const canManage = myMember?.role === "owner" || myMember?.role === "admin";
+  const activeTab = tabs.find((t) => t.id === tab) ?? tabs[0];
+  const lockedTab = TEAM_TABS.has(activeTab.id) && !isSynced;
+
+  return (
+    <SettingsModal
+      label={isSynced ? "Vault settings" : "Local vault settings"}
+      onClose={onClose}
+    >
+      <header className="settings-page-header">
+        <div className="settings-title">
+          <span className="settings-eyebrow">
+            {isSynced ? "Vault settings" : "Local vault"}
+          </span>
+          <h1>{activeOrg?.name ?? vault?.name ?? "Vault"}</h1>
+        </div>
+        <button className="icon-btn" onClick={onClose} aria-label="Close settings" title="Close (Esc)">
+          ✕
+        </button>
+      </header>
+
+      <div className="settings-body">
+        <nav className="settings-nav" aria-label="Settings sections">
+          {tabs.map((t) => {
+            const locked = TEAM_TABS.has(t.id) && !isSynced;
+            return (
+              <button
+                key={t.id}
+                type="button"
+                className={`menu-item${tab === t.id ? " active" : ""}${locked ? " locked" : ""}`}
+                onClick={() => setTab(t.id)}
+                title={locked ? "Turn on sync to unlock" : undefined}
+              >
+                {t.icon}
+                <span className="menu-item-label">{t.label}</span>
+                {locked && (
+                  <svg
+                    className="nav-lock"
+                    viewBox="0 0 24 24"
+                    fill="none"
+                    stroke="currentColor"
+                    strokeWidth="2"
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    aria-hidden="true"
+                  >
+                    <rect x="5" y="11" width="14" height="10" rx="2" />
+                    <path d="M8 11V7a4 4 0 0 1 8 0v4" />
+                  </svg>
+                )}
+              </button>
+            );
+          })}
+        </nav>
+
+        <section className="settings-content" aria-label={activeTab.label}>
+          <h2 className="settings-section-title">{activeTab.label}</h2>
+          {tab === "general" ? (
+            <GeneralTab
+              isSynced={isSynced}
+              canManage={canManage}
+              activeOrgName={activeOrg?.name ?? null}
+              onRequestSignIn={onRequestSignIn}
+            />
+          ) : lockedTab ? (
+            <SyncGate label={activeTab.label} onGoToSync={() => setTab("general")} />
+          ) : tab === "vaults" ? (
+            <VaultsTab />
+          ) : tab === "members" ? (
+            <MembersTab canManage={canManage} />
+          ) : tab === "billing" ? (
+            <BillingTab canManage={canManage} isSynced={isSynced} />
+          ) : tab === "access" ? (
+            <AccessPanel canManage={canManage} />
+          ) : tab === "mcp" ? (
+            <McpTab />
+          ) : tab === "versioning" ? (
+            <VersioningTab canManage={canManage} />
+          ) : tab === "import-export" ? (
+            <ImportExportTab />
+          ) : tab === "updates" ? (
+            <UpdatesTab />
+          ) : (
+            <AppearanceTab />
+          )}
+        </section>
+      </div>
+    </SettingsModal>
+  );
+}
+
+/**
+ * General tab: the identity of the current vault. Its heart is the
+ * Turn-on-sync card for a local vault — which adopts the folder you're
+ * already in (files and all) rather than making you set up a new one.
+ */
+function GeneralTab({
+  isSynced,
+  canManage,
+  activeOrgName,
+  onRequestSignIn,
+}: {
+  isSynced: boolean;
+  /** Owner/admin — the only roles that may flip a vault-wide latch. */
+  canManage: boolean;
+  activeOrgName: string | null;
+  onRequestSignIn?: () => void;
+}) {
+  const vault = useStore((s) => s.vault);
+  const syncStatus = useStore((s) => s.syncStatus);
+  const syncEnabledState = useStore((s) => s.syncEnabled);
+  const lastSyncedAt = useStore((s) => s.lastSyncedAt);
+  const syncPending = useStore((s) => s.syncPending);
+  // The vault's bulk-run counter ("Syncing 128/500"), so this row reports the
+  // whole vault's state and not just whether a socket is up.
+  const syncProgress = useStore((s) => s.syncProgress);
+  const serverUrl = useStore((s) => s.serverUrl);
+  const authStatus = useStore((s) => s.authStatus);
+
+  const [name, setName] = useState(vault?.name ?? "");
+  const [busy, setBusy] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+  const [limitNudge, setLimitNudge] = useState<{ kind: LimitKind; limit: number | null } | null>(
+    null,
+  );
+  const [upgradeOpen, setUpgradeOpen] = useState(false);
+
+  const turnOn = async () => {
+    if (authStatus !== "signed-in") {
+      // Not signed in yet — launch sign-in right here instead of dead-ending.
+      // After sign-in the button becomes "Turn on sync" (the page stays open).
+      onRequestSignIn?.();
+      return;
+    }
+    setBusy(true);
+    setError(null);
+    setLimitNudge(null);
+    try {
+      await useStore.getState().turnOnSyncForCurrentVault(name.trim() || undefined);
+    } catch (e) {
+      const kind = classifyLimitError(e);
+      if (kind) setLimitNudge({ kind, limit: limitFromError(e) });
+      else {
+        const message = e instanceof Error ? e.message : String(e);
+        setError(message);
+        // The inline line can be scrolled out of view, and the button simply
+        // returns to idle — which reads as "nothing happened". A sticky toast is
+        // always visible (#85).
+        toast(`Couldn't turn on sync — ${message}`, "error");
+      }
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  return (
+    <>
+      {isSynced ? (
+        <>
+          <div className="muted">
+            This vault syncs to your team. Its notes stay as plain files on
+            disk and live-sync to everyone with access.
+          </div>
+          <div className="menu-row">
+            <span className="menu-row-label">Sync</span>
+            <SyncBadge
+              status={syncStatus}
+              enabled={syncEnabledState}
+              lastSyncedAt={lastSyncedAt}
+              pending={syncPending}
+              progress={syncProgress}
+            />
+          </div>
+          <div className="menu-row">
+            <span className="menu-row-label">Server</span>
+            <code className="vault-root-path" title={serverUrl}>
+              {serverUrl}
+            </code>
+          </div>
+        </>
+      ) : (
+        <>
+          <div className="muted">
+            {activeOrgName
+              ? "You're viewing a local folder. Turn on sync to keep this vault on your account and across devices."
+              : "This vault lives only on this computer. Turn on sync to reach it from other devices — or invite people to it."}
+          </div>
+
+          <div className="sync-promo">
+            <h3 className="sync-promo-title">Turn on sync &amp; sharing</h3>
+            <p className="sync-promo-desc">
+              Keeps the notes and folders already here — nothing to re-import.
+              Enables live collaboration and lets you invite people. Private by
+              default; you choose what to share.
+            </p>
+            <div className="row invite-bar">
+              <input
+                placeholder="Vault name"
+                value={name}
+                onChange={(e) => setName(e.target.value)}
+                onKeyDown={(e) => {
+                  if (e.key === "Enter") void turnOn();
+                }}
+              />
+              <button className="primary" disabled={busy} onClick={() => void turnOn()}>
+                {busy
+                  ? "…"
+                  : authStatus === "signed-in"
+                    ? "Turn on sync"
+                    : "Sign in to turn on"}
+              </button>
+            </div>
+            {authStatus !== "signed-in" && (
+              <div className="muted">You'll need to sign in first — this button will prompt you.</div>
+            )}
+            {error && <div className="auth-error">{error}</div>}
+            {limitNudge && (
+              <LimitNudge
+                kind={limitNudge.kind}
+                limit={limitNudge.limit}
+                onUpgrade={() => setUpgradeOpen(true)}
+              />
+            )}
+          </div>
+        </>
+      )}
+
+      {isSynced && <FreezeRootRow canManage={canManage} />}
+
+      <div className="menu-sep" />
+      <div className="subhead">Folder on disk</div>
+      <div className="join-code-row">
+        <code className="vault-root-path" title={vault?.path ?? ""}>
+          {vault?.path ?? "—"}
+        </code>
+      </div>
+
+      {upgradeOpen && <UpgradeDialog onClose={() => setUpgradeOpen(false)} />}
+    </>
+  );
+}
+
+/**
+ * The "Freeze vault root" latch.
+ *
+ * A vault's top level is the one place where a stray note or folder is most
+ * visible and least recoverable — everyone sees it, and nobody is sure whose it
+ * is. Once a team has agreed the top-level shape, this closes it: new notes and
+ * folders have to go inside an existing folder.
+ *
+ * Deliberately applies to EVERYONE, owners and admins included, because the
+ * accidental root folder is almost always created by someone who does have
+ * permission. Only an owner/admin can lift it; everyone else sees the switch in
+ * its real state, disabled, so the rule is visible rather than mysterious.
+ */
+function FreezeRootRow({ canManage }: { canManage: boolean }) {
+  const rootFrozen = useStore((s) => s.rootFrozen);
+  const [busy, setBusy] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const flip = async (next: boolean) => {
+    setBusy(true);
+    setError(null);
+    try {
+      await useStore.getState().setRootFrozen(next);
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  return (
+    <>
+      <div className="menu-sep" />
+      <div className="subhead">Vault structure</div>
+      <label className="menu-row toggle-row">
+        <span className="menu-row-label">
+          Freeze vault root
+          <span className="field-hint">
+            Stops anything new being created at the top level of this vault —
+            new notes and folders have to go inside an existing folder. Applies
+            to everyone, including you; only an owner or admin can turn it off.
+            Nothing already at the root is moved, renamed, or hidden.
+          </span>
+        </span>
+        <Switch
+          checked={rootFrozen}
+          disabled={!canManage || busy}
+          ariaLabel="Freeze vault root"
+          title={canManage ? undefined : "Only an owner or admin can change this"}
+          onChange={(next) => void flip(next)}
+        />
+      </label>
+      {!canManage && (
+        <div className="muted">
+          {rootFrozen
+            ? "This vault's root is frozen. Ask an owner or admin to unfreeze it."
+            : "Only an owner or admin can freeze this vault's root."}
+        </div>
+      )}
+      {error && <div className="auth-error">{error}</div>}
+    </>
+  );
+}
+
+/** Locked-section gate shown for a team tab on a local vault. */
+function SyncGate({ label, onGoToSync }: { label: string; onGoToSync: () => void }) {
+  return (
+    <div className="sync-gate">
+      <svg
+        className="sync-gate-icon"
+        viewBox="0 0 24 24"
+        fill="none"
+        stroke="currentColor"
+        strokeWidth="1.6"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        aria-hidden="true"
+      >
+        <rect x="4" y="11" width="16" height="10" rx="2" />
+        <path d="M8 11V7a4 4 0 0 1 8 0v4" />
+      </svg>
+      <h3>{label} unlocks with sync</h3>
+      <p className="muted">
+        Turn on sync for this vault to invite people, set sharing and
+        permissions, and connect AI clients.
+      </p>
+      <button className="primary" onClick={onGoToSync}>
+        Turn on sync &amp; sharing →
+      </button>
+    </div>
+  );
+}
+
+/**
+ * Vaults: switch between vaults, create/join, and manage where their
+ * local folders live. Each vault owns one folder under the managed root;
+ * switching swaps the sidebar to that vault's folder and repoints the
+ * stable `current` symlink external tools point at.
+ */
+function VaultsTab() {
+  const session = useStore((s) => s.session);
+  const organizations = useStore((s) => s.organizations);
+  const members = useStore((s) => s.members);
+  const vault = useStore((s) => s.vault);
+  const syncEnabled = useStore((s) => s.syncEnabled);
+  const billingEnabled = useStore((s) => s.billingConfig?.enabled === true);
+  // Bumped after a local remove/delete so the recents list re-fetches.
+  const [localsNonce, setLocalsNonce] = useState(0);
+  const locals = useLocalVaults(localsNonce);
+  const knownOrgIds = useKnownOrgIds();
+
+  const [root, setRoot] = useState<string | null>(null);
+  const [bound, setBound] = useState<Record<string, string>>(() => readOrgVaults());
+  const [creating, setCreating] = useState(false);
+  const [orgName, setOrgName] = useState("");
+  const [joining, setJoining] = useState(false);
+  const [joinCode, setJoinCode] = useState("");
+  const [joinError, setJoinError] = useState<string | null>(null);
+  const [busy, setBusy] = useState(false);
+  // orgId whose permanent deletion is awaiting a second confirming click.
+  const [confirmDelete, setConfirmDelete] = useState<string | null>(null);
+  // A vault that is actually PAYING, whose deletion needs the full dialog
+  // instead: the two-click row has nowhere to say what happens to the money
+  // (#111). Its billing snapshot rides along so the copy can name the date.
+  const [subDelete, setSubDelete] = useState<
+    { orgId: string; name: string; billing: OrgBilling } | null
+  >(null);
+  // local-vault path whose file deletion is awaiting a second confirming click.
+  const [confirmDeleteLocal, setConfirmDeleteLocal] = useState<string | null>(null);
+  // A vault the user is about to leave (#121). Always the full dialog: it has
+  // to say that the folder on this device goes too, which a row can't.
+  const [confirmLeave, setConfirmLeave] = useState<{ orgId: string; name: string } | null>(
+    null,
+  );
+  const [actionError, setActionError] = useState<string | null>(null);
+  // Free-plan vault-cap hit while creating — shows an upgrade nudge instead.
+  const [limitNudge, setLimitNudge] = useState<{ kind: LimitKind; limit: number | null } | null>(
+    null,
+  );
+  const [upgradeOpen, setUpgradeOpen] = useState(false);
+
+  useEffect(() => {
+    let cancelled = false;
+    ipc
+      .getVaultsRoot()
+      .then((r) => {
+        if (!cancelled) setRoot(r);
+      })
+      .catch(() => {});
+    return () => {
+      cancelled = true;
+    };
+  }, []);
+
+  const activeOrgId = session?.activeOrganizationId ?? null;
+  // We only know the caller's role for the ACTIVE vault (members are loaded
+  // for it alone). On the active row we can therefore hide Delete from
+  // non-owners; on other rows we can't tell, so we show it and let the server
+  // enforce owner-only (403, surfaced via actionError). `deleteRemoteVault` takes
+  // an explicit org id, so deleting a non-active vault works without first
+  // switching to it.
+  const isActiveOwner =
+    members.find((m) => m.userId === session?.user.id)?.role === "owner";
+  const canDelete = (orgId: string) =>
+    orgId === activeOrgId ? isActiveOwner : true;
+  // Same uncertainty, mirrored: on the active row Leave is for non-owners
+  // only; elsewhere both are offered and the server's 409 settles it.
+  const canLeave = (orgId: string) =>
+    orgId === activeOrgId ? !isActiveOwner : true;
+
+  const folderName = (orgId: string): string | null => {
+    const p = bound[orgId];
+    return p ? (p.split("/").pop() ?? p) : null;
+  };
+
+  // The org whose folder is actually open now — the true "Current", vs. merely
+  // the account's active org (you can be viewing a local folder with sync off).
+  const openPath = vault?.path ?? null;
+  const isOpenOrg = (orgId: string) => openPath != null && bound[orgId] === openPath;
+
+  const switchTo = async (orgId: string) => {
+    if (busy || isOpenOrg(orgId)) return;
+    setBusy(true);
+    try {
+      await useStore.getState().setActiveOrganization(orgId);
+      setBound(readOrgVaults());
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const switchToLocal = async (path: string) => {
+    if (busy) return;
+    setBusy(true);
+    try {
+      await useStore.getState().openLocalVault(path);
+      setBound(readOrgVaults());
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  // Detach a vault from this device only (server data untouched).
+  const removeLocal = async (orgId: string) => {
+    if (busy) return;
+    setBusy(true);
+    setActionError(null);
+    try {
+      await useStore.getState().removeVaultLocally(orgId);
+      setBound(readOrgVaults());
+    } catch (e) {
+      setActionError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  /**
+   * Ask before deleting. A vault that is actually paying gets the full
+   * ConfirmDialog, because "Delete everything?" cannot say the one thing its
+   * owner needs to know: the subscription stops at the END of the current
+   * period, and until then it can be moved to another vault (#111). Every
+   * other vault keeps today's two-click row.
+   *
+   * A billing lookup that fails is treated as free — an unreachable billing
+   * endpoint must not block a delete the user is entitled to make.
+   */
+  const askDelete = async (orgId: string, name: string) => {
+    setActionError(null);
+    if (!billingEnabled) {
+      setConfirmDelete(orgId);
+      return;
+    }
+    let billing: OrgBilling | null = null;
+    try {
+      billing = await authManager.api.getOrgBilling(orgId);
+    } catch {
+      billing = null;
+    }
+    if (billing && (billing.status === "active" || billing.status === "past_due")) {
+      setSubDelete({ orgId, name, billing });
+    } else {
+      setConfirmDelete(orgId);
+    }
+  };
+
+  // Permanently delete a vault everywhere (owner only, confirmed above).
+  const deletePermanently = async (orgId: string) => {
+    if (busy) return;
+    setBusy(true);
+    setActionError(null);
+    try {
+      const result = await useStore.getState().deleteRemoteVault(orgId);
+      setBound(readOrgVaults());
+      setConfirmDelete(null);
+      setSubDelete(null);
+      if (result.subscription) {
+        // Neutral, not success: the vault is gone, but the user is still paying
+        // for the rest of the period and that time is recoverable.
+        const ends = result.subscription.currentPeriodEnd;
+        toast(
+          ends
+            ? `Vault deleted. Pro ends on ${formatDate(ends)} — move it from Billing if you want to keep it.`
+            : "Vault deleted. Pro ends when the current period does — move it from Billing if you want to keep it.",
+          "neutral",
+        );
+      }
+    } catch (e) {
+      const message = e instanceof Error ? e.message : String(e);
+      // Covers the 502 `subscription_cancel_failed` case: the server's message
+      // rides `ApiError.message`, and NOTHING was deleted. The dialog stays
+      // open (only success clears it) so the error has somewhere to show.
+      setActionError(message);
+      // Destructive path: a failure here must never look like a success (#85).
+      toast(`Couldn't delete the vault — ${message}`, "error");
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  // Leave a vault someone else owns (confirmed above). Server first, then the
+  // vault leaves this device entirely; the dialog stays open on failure so the
+  // server's reason (an owner's 409, offline) has somewhere to show.
+  const leaveVault = async (orgId: string, name: string) => {
+    if (busy) return;
+    setBusy(true);
+    setActionError(null);
+    try {
+      await useStore.getState().leaveVault(orgId);
+      setBound(readOrgVaults());
+      setConfirmLeave(null);
+      toast(`You left ${name}.`, "neutral");
+    } catch (e) {
+      const message = e instanceof Error ? e.message : String(e);
+      setActionError(message);
+      toast(`Couldn't leave the vault — ${message}`, "error");
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  // Forget a local vault from this device's list (files on disk are kept).
+  const removeLocalVaultRow = async (path: string) => {
+    if (busy) return;
+    setBusy(true);
+    setActionError(null);
+    try {
+      await useStore.getState().removeLocalVault(path);
+      setLocalsNonce((n) => n + 1);
+    } catch (e) {
+      setActionError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  // Move a local vault's folder to the OS trash (destructive, two-click confirm).
+  const deleteLocalFiles = async (path: string) => {
+    if (busy) return;
+    setBusy(true);
+    setActionError(null);
+    try {
+      await useStore.getState().deleteLocalVault(path);
+      setConfirmDeleteLocal(null);
+      setLocalsNonce((n) => n + 1);
+    } catch (e) {
+      setActionError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const createOrg = async () => {
+    if (!orgName.trim()) return;
+    setBusy(true);
+    setActionError(null);
+    setLimitNudge(null);
+    try {
+      await useStore.getState().createOrganization(orgName.trim());
+      setOrgName("");
+      setCreating(false);
+      setBound(readOrgVaults());
+    } catch (e) {
+      // A 402 vault-cap rejection becomes an upgrade nudge; anything else is
+      // a real error (previously swallowed silently — that was the create bug).
+      const kind = classifyLimitError(e);
+      if (kind) setLimitNudge({ kind, limit: limitFromError(e) });
+      else setActionError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const joinByCode = async () => {
+    if (!joinCode.trim()) return;
+    setBusy(true);
+    setJoinError(null);
+    try {
+      await useStore.getState().joinVault(joinCode);
+      setJoinCode("");
+      setJoining(false);
+      setBound(readOrgVaults());
+    } catch (e) {
+      setJoinError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const changeRoot = async () => {
+    try {
+      const picked = await ipc.pickVaultsRoot();
+      if (picked) setRoot(picked);
+    } catch {
+      /* picker cancelled/unavailable */
+    }
+  };
+
+  // Active vault pinned to the top.
+  const ordered = [
+    ...organizations.filter((o) => o.id === activeOrgId),
+    ...organizations.filter((o) => o.id !== activeOrgId),
+  ];
+
+  // A folder still carrying a binding to a vault this account doesn't have:
+  // the vault was deleted, or belongs to another account, or the binding is
+  // just stale. It's listed here (not silently swallowed as "bound") so the
+  // user's notes are never on disk and absent from every list at once.
+  const orphanBinding = (path: string) =>
+    Object.entries(bound).some(([orgId, p]) => p === path && !knownOrgIds.has(orgId));
+
+  const localsOrdered = [
+    ...locals.filter((r) => !syncEnabled && vault?.path === r.path),
+    ...locals.filter((r) => !(!syncEnabled && vault?.path === r.path)),
+  ];
+
+  return (
+    <>
+      {session && (
+        <>
+      <div className="subhead">In this account ({organizations.length})</div>
+      <ul className="member-list vault-list">
+        {ordered.map((o) => {
+          const isActive = isOpenOrg(o.id);
+          const fname = folderName(o.id);
+          return (
+            <li key={o.id}>
+              <span className="menu-swatch" aria-hidden="true">
+                {o.name[0]?.toUpperCase() ?? "?"}
+              </span>
+              <span className="member-name">
+                {o.name}
+                <span className="muted vault-folder">
+                  {" "}
+                  {fname ? `· ${fname}` : "· folder created on first open"}
+                </span>
+              </span>
+              {confirmDelete === o.id ? (
+                <span className="vault-row-actions">
+                  <span className="muted">Delete everything?</span>
+                  <button
+                    className="link-btn"
+                    disabled={busy}
+                    onClick={() => setConfirmDelete(null)}
+                  >
+                    Cancel
+                  </button>
+                  <AsyncButton
+                    className="link-btn danger"
+                    disabled={busy}
+                    onClick={() => deletePermanently(o.id)}
+                  >
+                    Delete
+                  </AsyncButton>
+                </span>
+              ) : (
+                <span className="vault-row-actions">
+                  {isActive ? (
+                    <span className="member-role">Current</span>
+                  ) : (
+                    <AsyncButton
+                      className="link-btn"
+                      disabled={busy}
+                      onClick={() => switchTo(o.id)}
+                    >
+                      Switch
+                    </AsyncButton>
+                  )}
+                  <AsyncButton
+                    className="link-btn"
+                    disabled={busy}
+                    title="Stop syncing this vault here; server data is kept"
+                    onClick={() => removeLocal(o.id)}
+                  >
+                    Remove from device
+                  </AsyncButton>
+                  {canLeave(o.id) && (
+                    <button
+                      className="link-btn danger"
+                      disabled={busy}
+                      title="Leave this vault — you lose access and it is removed from this device"
+                      onClick={() => {
+                        setActionError(null);
+                        setConfirmLeave({ orgId: o.id, name: o.name });
+                      }}
+                    >
+                      Leave
+                    </button>
+                  )}
+                  {canDelete(o.id) && (
+                    <AsyncButton
+                      className="link-btn danger"
+                      disabled={busy}
+                      title="Permanently delete this vault and all its notes for everyone"
+                      onClick={() => askDelete(o.id, o.name)}
+                    >
+                      Delete
+                    </AsyncButton>
+                  )}
+                </span>
+              )}
+            </li>
+          );
+        })}
+      </ul>
+
+      <div className="vault-tab-actions">
+        {creating ? (
+          <form
+            className="vault-tab-form"
+            onSubmit={(e) => {
+              e.preventDefault();
+              void createOrg();
+            }}
+          >
+            <input
+              autoFocus
+              placeholder="Vault name"
+              value={orgName}
+              onChange={(e) => setOrgName(e.target.value)}
+              onKeyDown={(e) => {
+                if (e.key === "Escape") setCreating(false);
+              }}
+            />
+            <button
+              type="submit"
+              className="primary"
+              disabled={busy || !orgName.trim()}
+            >
+              Create
+            </button>
+            <button
+              type="button"
+              className="ghost-pill"
+              disabled={busy}
+              onClick={() => setCreating(false)}
+            >
+              Cancel
+            </button>
+          </form>
+        ) : joining ? (
+          <form
+            className="vault-tab-form"
+            onSubmit={(e) => {
+              e.preventDefault();
+              void joinByCode();
+            }}
+          >
+            <input
+              autoFocus
+              className="vault-tab-code"
+              placeholder="Join code, e.g. K7MPX2RA"
+              value={joinCode}
+              spellCheck={false}
+              autoCapitalize="characters"
+              onChange={(e) => setJoinCode(e.target.value.toUpperCase())}
+              onKeyDown={(e) => {
+                if (e.key === "Escape") setJoining(false);
+              }}
+            />
+            <button
+              type="submit"
+              className="primary"
+              disabled={busy || !joinCode.trim()}
+            >
+              Join
+            </button>
+            <button
+              type="button"
+              className="ghost-pill"
+              disabled={busy}
+              onClick={() => setJoining(false)}
+            >
+              Cancel
+            </button>
+          </form>
+        ) : (
+          <>
+            <button
+              type="button"
+              className="ghost-pill vault-tab-add"
+              onClick={() => setCreating(true)}
+            >
+              <svg viewBox="0 0 24 24" width="15" height="15" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+                <path d="M12 5v14M5 12h14" />
+              </svg>
+              <span>New vault</span>
+            </button>
+            <button
+              type="button"
+              className="ghost-pill vault-tab-add"
+              onClick={() => setJoining(true)}
+            >
+              <svg viewBox="0 0 24 24" width="15" height="15" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+                <path d="M4 9h16M4 15h16M10 3 8 21M16 3l-2 18" />
+              </svg>
+              <span>Join with code</span>
+            </button>
+          </>
+        )}
+      </div>
+      {joinError && <div className="auth-error">{joinError}</div>}
+      {actionError && <div className="auth-error">{actionError}</div>}
+      {limitNudge && (
+        <LimitNudge
+          kind={limitNudge.kind}
+          limit={limitNudge.limit}
+          onUpgrade={() => setUpgradeOpen(true)}
+        />
+      )}
+        </>
+      )}
+
+      {localsOrdered.length > 0 && (
+        <>
+          {session && <div className="menu-sep" />}
+          <div className="subhead">On this device ({localsOrdered.length})</div>
+          <div className="muted">
+            Local folders open on this computer. They aren't on your account —
+            turn on sync from a vault's settings to reach it elsewhere.
+          </div>
+          <ul className="member-list vault-list">
+            {localsOrdered.map((r) => {
+              const isCurrent = !syncEnabled && vault?.path === r.path;
+              return (
+                <li key={r.path}>
+                  <span className="menu-swatch" aria-hidden="true">
+                    {r.name[0]?.toUpperCase() ?? "?"}
+                  </span>
+                  <span className="member-name">
+                    {r.name}
+                    <span
+                      className="muted vault-folder"
+                      title={
+                        orphanBinding(r.path)
+                          ? `${r.path} — this folder was synced to a vault that isn't in this account`
+                          : r.path
+                      }
+                    >
+                      {orphanBinding(r.path) ? " · Local · was synced elsewhere" : " · Local"}
+                    </span>
+                  </span>
+                  {confirmDeleteLocal === r.path ? (
+                    <span className="vault-row-actions">
+                      <span className="muted">Delete this vault?</span>
+                      <button
+                        className="link-btn"
+                        disabled={busy}
+                        onClick={() => setConfirmDeleteLocal(null)}
+                      >
+                        Cancel
+                      </button>
+                      <AsyncButton
+                        className="link-btn danger"
+                        disabled={busy}
+                        onClick={() => deleteLocalFiles(r.path)}
+                      >
+                        Delete
+                      </AsyncButton>
+                    </span>
+                  ) : (
+                    <span className="vault-row-actions">
+                      {isCurrent ? (
+                        <span className="member-role">Current</span>
+                      ) : (
+                        <AsyncButton
+                          className="link-btn"
+                          disabled={busy}
+                          onClick={() => switchToLocal(r.path)}
+                        >
+                          Switch
+                        </AsyncButton>
+                      )}
+                      <button
+                        className="link-btn"
+                        disabled={busy}
+                        title="Remove this folder from the list. Files stay on disk."
+                        onClick={() => void removeLocalVaultRow(r.path)}
+                      >
+                        Remove
+                      </button>
+                      <button
+                        className="link-btn danger"
+                        disabled={busy}
+                        title="Delete this vault — moves its folder and all its notes to the Trash"
+                        onClick={() => {
+                          setActionError(null);
+                          setConfirmDeleteLocal(r.path);
+                        }}
+                      >
+                        Delete
+                      </button>
+                    </span>
+                  )}
+                </li>
+              );
+            })}
+          </ul>
+        </>
+      )}
+
+      {session && (
+        <>
+          <div className="menu-sep" />
+          <div className="subhead">Vault folder location</div>
+          <div className="muted">
+            New vaults get their own folder here. The active vault is also
+            linked at <code>current</code> so tools like Claude Desktop can point at
+            one fixed path.
+          </div>
+          <div className="join-code-row">
+            <code className="vault-root-path" title={root ?? ""}>
+              {root ?? "…"}
+            </code>
+            <button className="link-btn" onClick={() => void changeRoot()}>
+              Change…
+            </button>
+          </div>
+        </>
+      )}
+
+      {upgradeOpen && <UpgradeDialog onClose={() => setUpgradeOpen(false)} />}
+
+      {confirmLeave && (
+        <ConfirmDialog
+          title={`Leave ${confirmLeave.name}?`}
+          confirmLabel="Leave vault"
+          onCancel={() => setConfirmLeave(null)}
+          onConfirm={() => leaveVault(confirmLeave.orgId, confirmLeave.name)}
+        >
+          <p>
+            You lose access to this vault on all your devices right away, and the
+            owner is told that you left.
+          </p>
+          <p>
+            {bound[confirmLeave.orgId] ? (
+              <>
+                Its folder on this device, <strong>{folderName(confirmLeave.orgId)}</strong>,
+                moves to the Trash.
+              </>
+            ) : (
+              "Nothing from it is stored on this device."
+            )}{" "}
+            The vault itself and everyone else's access are unchanged.
+          </p>
+          <p>To come back later, you'll need a new invitation or join code.</p>
+          {actionError && <div className="auth-error">{actionError}</div>}
+        </ConfirmDialog>
+      )}
+
+      {subDelete && (
+        <ConfirmDialog
+          title={`Delete ${subDelete.name}?`}
+          confirmLabel="Delete vault"
+          onCancel={() => setSubDelete(null)}
+          onConfirm={() => deletePermanently(subDelete.orgId)}
+        >
+          <p>
+            This vault is on <strong>Pro</strong>
+            {subDelete.billing.currentPeriodEnd
+              ? subDelete.billing.cancelAtPeriodEnd
+                ? `, ending ${formatDate(subDelete.billing.currentPeriodEnd)}`
+                : `, renewing ${formatDate(subDelete.billing.currentPeriodEnd)}`
+              : ""}
+            .
+          </p>
+          <p>
+            Deleting it stops the subscription at the end of the current period.
+            You won't be charged again, and until then you can move the
+            subscription to another vault from <strong>Billing</strong>.
+          </p>
+          <p>
+            Every note, folder and attachment in this vault is deleted for
+            everyone. That part can't be undone.
+          </p>
+          {actionError && <div className="auth-error">{actionError}</div>}
+        </ConfirmDialog>
+      )}
+    </>
+  );
+}
+
+function MembersTab({ canManage }: { canManage: boolean }) {
+  const session = useStore((s) => s.session);
+  const members = useStore((s) => s.members);
+  const pendingInvitations = useStore((s) => s.pendingInvitations);
+  // Invite links are built against the server this vault lives on, not a
+  // constant: a self-hosted vault's invitation only resolves on its own server.
+  const serverUrl = useStore((s) => s.serverUrl);
+
+  const [inviteEmail, setInviteEmail] = useState("");
+  const [inviteRole, setInviteRole] = useState<"member" | "admin">("member");
+  const [busy, setBusy] = useState(false);
+  const [code, setCode] = useState<string | null>(null);
+  const [copied, setCopied] = useState(false);
+  // The invitation just created, so its link can be shown. Not read out of
+  // `pendingInvitations`: that list is keyed by email and makes no promise
+  // about which row is the one this click produced.
+  const [created, setCreated] = useState<InviteResult | null>(null);
+  // Which link was copied, by invitation id ("new" for the notice above the
+  // list) — one shared flag would tick every row at once.
+  const [copiedLink, setCopiedLink] = useState<string | null>(null);
+  // Revoking is destructive and unprompted-recoverable only by re-inviting, so
+  // it gets the same inline "Revoke → Confirm" as member removal.
+  const [confirmRevokeId, setConfirmRevokeId] = useState<string | null>(null);
+  // Invite errors had no home before — surface them here (silent-failure fix).
+  const [inviteError, setInviteError] = useState<string | null>(null);
+  const [limitNudge, setLimitNudge] = useState<{ kind: LimitKind; limit: number | null } | null>(
+    null,
+  );
+  const [upgradeOpen, setUpgradeOpen] = useState(false);
+  // Member removal: an inline "Remove → Confirm" per row so it's never one click.
+  const [confirmRemoveId, setConfirmRemoveId] = useState<string | null>(null);
+  const [removeBusy, setRemoveBusy] = useState(false);
+  const [removeError, setRemoveError] = useState<string | null>(null);
+
+  // Role changes: no confirm step (reversible, unlike remove), one busy row
+  // at a time so a slow server can't interleave two changes.
+  const [roleBusyId, setRoleBusyId] = useState<string | null>(null);
+  const [roleError, setRoleError] = useState<string | null>(null);
+
+  // The caller's own role in this vault, so we mirror the server's rules and
+  // only offer Remove / role changes where they would actually succeed
+  // (see memberRoles.ts for the shared matrix).
+  const myRole = members.find((m) => m.userId === session?.user.id)?.role;
+  const canAct = (m: Member): boolean =>
+    canActOnMember({
+      canManage,
+      myUserId: session?.user.id,
+      myRole,
+      target: { userId: m.userId, role: m.role },
+    });
+
+  const doRemove = async (userId: string) => {
+    setRemoveBusy(true);
+    setRemoveError(null);
+    try {
+      await useStore.getState().removeMember(userId);
+      setConfirmRemoveId(null);
+    } catch (e) {
+      setRemoveError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setRemoveBusy(false);
+    }
+  };
+
+  const doChangeRole = async (userId: string, role: "member" | "admin") => {
+    setRoleBusyId(userId);
+    setRoleError(null);
+    try {
+      await useStore.getState().updateMemberRole(userId, role);
+    } catch (e) {
+      setRoleError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setRoleBusyId(null);
+    }
+  };
+
+  // The vault's shareable join code (owner/admin only; server creates it
+  // lazily). Older servers without the endpoint simply hide the section.
+  useEffect(() => {
+    if (!canManage) return;
+    let cancelled = false;
+    authManager.api
+      .getJoinCode()
+      .then((c) => {
+        if (!cancelled) setCode(c);
+      })
+      .catch(() => {});
+    return () => {
+      cancelled = true;
+    };
+  }, [canManage]);
+
+  const copyCode = async () => {
+    if (!code) return;
+    try {
+      await navigator.clipboard.writeText(code);
+      setCopied(true);
+      window.setTimeout(() => setCopied(false), 1500);
+    } catch {
+      /* clipboard unavailable */
+    }
+  };
+
+  /** Copy one invitation's https link. `tag` keys the "Copied ✓" to its row. */
+  const copyInviteLink = async (invitationId: string, tag: string) => {
+    const link = buildInviteLink(serverUrl, invitationId);
+    if (!link) return;
+    try {
+      await navigator.clipboard.writeText(link);
+      setCopiedLink(tag);
+      window.setTimeout(() => setCopiedLink((c) => (c === tag ? null : c)), 1500);
+    } catch {
+      /* clipboard unavailable */
+    }
+  };
+
+  const revoke = async (invitationId: string) => {
+    setInviteError(null);
+    try {
+      await authManager.api.cancelInvitation(invitationId);
+      setConfirmRevokeId(null);
+      // Drop the just-created notice if it was about this invitation — its link
+      // is dead now, and offering to copy it would be worse than saying nothing.
+      setCreated((c) => (c?.invitation.id === invitationId ? null : c));
+      await useStore.getState().refreshVault();
+    } catch (e) {
+      setInviteError(e instanceof Error ? e.message : String(e));
+    }
+  };
+
+  const invite = async () => {
+    if (!inviteEmail.trim()) return;
+    setBusy(true);
+    setInviteError(null);
+    setLimitNudge(null);
+    try {
+      setCreated(await useStore.getState().inviteMember(inviteEmail.trim(), inviteRole));
+      setInviteEmail("");
+    } catch (e) {
+      // A 402 member-cap rejection becomes an upgrade nudge; anything else is a
+      // real error (this tab had no error slot before — that was the bug).
+      const kind = classifyLimitError(e);
+      if (kind) setLimitNudge({ kind, limit: limitFromError(e) });
+      else setInviteError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  return (
+    <>
+      {canManage && code && (
+        <div className="join-code-row">
+          <div className="join-code-meta">
+            <span className="subhead">Join code</span>
+            <span className="muted">
+              Teammates pick “Join with code” in their account menu after signing in.
+              Someone who was also invited by email lands with the invited role either
+              way.
+            </span>
+          </div>
+          <code className="join-code">{code}</code>
+          <button className="link-btn" onClick={() => void copyCode()}>
+            {copied ? "Copied ✓" : "Copy"}
+          </button>
+        </div>
+      )}
+      {canManage && (
+        <div className="row invite-bar">
+          <input
+            type="email"
+            placeholder="email@team.com"
+            value={inviteEmail}
+            autoFocus
+            onChange={(e) => setInviteEmail(e.target.value)}
+            onKeyDown={(e) => {
+              if (e.key === "Enter") void invite();
+            }}
+          />
+          <RoleSelect
+            variant="field"
+            value={inviteRole}
+            onSelect={setInviteRole}
+            ariaLabel="Invite role"
+          />
+          <button className="primary" disabled={busy} onClick={() => void invite()}>
+            Invite
+          </button>
+        </div>
+      )}
+      {inviteError && <div className="auth-error">{inviteError}</div>}
+      {/* Inline, not a toast: this modal has its own error/notice slots, and a
+          corner toast carrying a link the admin has to COPY is a link they will
+          lose. */}
+      {created && (
+        <div className={`invite-notice${created.emailError ? " is-warning" : ""}`}>
+          {created.emailed ? (
+            // A fact, not a hope: the server only says so once the mail
+            // provider has accepted the message.
+            <span>Invitation emailed to {created.invitation.email}.</span>
+          ) : (
+            <>
+              <span>
+                {created.emailError
+                  ? `Invitation created, but the email to ${created.invitation.email} couldn't be sent (${created.emailError}). Share this link instead:`
+                  : `Invitation created — this server doesn't send email, so share this link with ${created.invitation.email}:`}
+              </span>
+              <div className="invite-notice-link">
+                <code>{buildInviteLink(serverUrl, created.invitation.id) ?? ""}</code>
+                <button
+                  className="link-btn"
+                  onClick={() => void copyInviteLink(created.invitation.id, "new")}
+                >
+                  {copiedLink === "new" ? "Copied ✓" : "Copy"}
+                </button>
+              </div>
+            </>
+          )}
+        </div>
+      )}
+      {limitNudge && (
+        <LimitNudge
+          kind={limitNudge.kind}
+          limit={limitNudge.limit}
+          onUpgrade={() => setUpgradeOpen(true)}
+        />
+      )}
+
+      <div className="subhead">In this vault ({members.length})</div>
+      <ul className="member-list">
+        {members.map((m) => {
+          const label = m.user?.name || m.user?.email || m.userId;
+          return (
+            <li key={m.id}>
+              <Avatar label={label} />
+              <span className="member-name">
+                {label}
+                {m.userId === session?.user.id && <span className="muted"> (you)</span>}
+              </span>
+              {canAct(m) && (m.role === "member" || m.role === "admin") ? (
+                <RoleSelect
+                  variant="pill"
+                  value={m.role}
+                  disabled={roleBusyId !== null}
+                  ariaLabel={`Change role of ${label}`}
+                  onSelect={(r) => void doChangeRole(m.userId, r)}
+                />
+              ) : (
+                <span className={`member-role ${m.role}`}>{m.role}</span>
+              )}
+              {canAct(m) &&
+                (confirmRemoveId === m.userId ? (
+                  <>
+                    <AsyncButton
+                      className="link-btn danger"
+                      disabled={removeBusy}
+                      onClick={() => doRemove(m.userId)}
+                    >
+                      Confirm
+                    </AsyncButton>
+                    <button
+                      className="link-btn"
+                      disabled={removeBusy}
+                      onClick={() => setConfirmRemoveId(null)}
+                    >
+                      Cancel
+                    </button>
+                  </>
+                ) : (
+                  <button
+                    className="link-btn danger"
+                    onClick={() => {
+                      setRemoveError(null);
+                      setConfirmRemoveId(m.userId);
+                    }}
+                  >
+                    Remove
+                  </button>
+                ))}
+            </li>
+          );
+        })}
+      </ul>
+      {removeError && <div className="auth-error">{removeError}</div>}
+      {roleError && <div className="auth-error">{roleError}</div>}
+
+      {pendingInvitations.length > 0 && (
+        <>
+          <div className="subhead">Invited — awaiting response</div>
+          <ul className="member-list">
+            {pendingInvitations.map((inv) => (
+              <li key={inv.id}>
+                <Avatar label={inv.email} />
+                <span className="member-name">{inv.email}</span>
+                <span className="member-role pending">{inv.role} · pending</span>
+                {canManage && (
+                  <>
+                    {/* The link is useful long after the invite was sent: the
+                        email may have bounced, or this server may not send any. */}
+                    <button
+                      className="link-btn"
+                      onClick={() => void copyInviteLink(inv.id, inv.id)}
+                    >
+                      {copiedLink === inv.id ? "Copied ✓" : "Copy link"}
+                    </button>
+                    {confirmRevokeId === inv.id ? (
+                      <>
+                        <AsyncButton
+                          className="link-btn danger"
+                          onClick={() => revoke(inv.id)}
+                        >
+                          Confirm
+                        </AsyncButton>
+                        <button
+                          className="link-btn"
+                          onClick={() => setConfirmRevokeId(null)}
+                        >
+                          Cancel
+                        </button>
+                      </>
+                    ) : (
+                      <button
+                        className="link-btn danger"
+                        onClick={() => {
+                          setInviteError(null);
+                          setConfirmRevokeId(inv.id);
+                        }}
+                      >
+                        Revoke
+                      </button>
+                    )}
+                  </>
+                )}
+              </li>
+            ))}
+          </ul>
+        </>
+      )}
+
+      {upgradeOpen && <UpgradeDialog onClose={() => setUpgradeOpen(false)} />}
+    </>
+  );
+}
+
+/**
+ * BillingTab: the current vault's plan and seats, then every subscription the
+ * signed-in user can act on (spec 04, #109/#110).
+ *
+ * Three sections, in the order someone reaching for this page needs them:
+ * this vault (what am I on?), all my vaults (what am I paying for?), and
+ * subscriptions from deleted vaults (what am I paying for that no longer
+ * exists?). Facts are visible to every member; each action is gated to the
+ * role the server will actually accept — Upgrade and Manage for owners and
+ * admins, Transfer for owners only, because it changes who pays for what.
+ *
+ * Only the first section needs a synced vault; the other two are account-wide,
+ * which is why this tab is no longer in TEAM_TABS.
+ */
+function BillingTab({ canManage, isSynced }: { canManage: boolean; isSynced: boolean }) {
+  const billingConfig = useStore((s) => s.billingConfig);
+  const orgBilling = useStore((s) => s.orgBilling);
+  const myBilling = useStore((s) => s.myBilling);
+  const orgId = useStore((s) => s.session?.activeOrganizationId ?? null);
+
+  // The vault the upgrade dialog should charge: the active one from the card
+  // at the top, or any owned free vault picked out of the list below.
+  const [upgradeOrg, setUpgradeOrg] = useState<string | null>(null);
+  const [busy, setBusy] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+  // A transfer being set up: the source row's Transfer was clicked and the
+  // dialog is open. `targetOrgId` is the pick, null until one is made (or the
+  // only eligible vault, pre-picked). It moves money between vaults, so the
+  // dialog is where both the choice AND the confirmation happen — never a
+  // popover that fires on the first click.
+  const [transfer, setTransfer] = useState<{
+    sourceOrgId: string;
+    sourceLabel: string;
+    targetOrgId: string | null;
+  } | null>(null);
+  // An orphaned subscription waiting on a "cancel now" confirmation.
+  const [cancelling, setCancelling] = useState<{ orgId: string; label: string } | null>(
+    null,
+  );
+
+  // Refresh this vault's seats AND the account-wide list whenever the tab opens.
+  useEffect(() => {
+    void useStore.getState().refreshOrgBilling();
+    void useStore.getState().refreshMyBilling();
+  }, []);
+
+  const refreshAll = async () => {
+    await Promise.all([
+      useStore.getState().refreshMyBilling(),
+      useStore.getState().refreshOrgBilling(),
+    ]);
+  };
+
+  /** Open a vault's provider portal in the OS browser (owner/admin). */
+  const openPortal = async (portalOrgId: string) => {
+    if (busy) return;
+    setBusy(true);
+    setError(null);
+    try {
+      const { url } = await authManager.api.getBillingPortalUrl(portalOrgId);
+      await ipc.openExternal(url);
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  // Both confirmed actions leave their dialog OPEN on failure and report into
+  // it, rather than closing over an error nobody sees. Neither re-throws — the
+  // dialog's own AsyncButton has finished reporting by then.
+  const runTransfer = async () => {
+    if (!transfer || !transfer.targetOrgId) return;
+    const { sourceOrgId, targetOrgId } = transfer;
+    const targetName = vaults.find((v) => v.orgId === targetOrgId)?.name ?? "the vault";
+    setError(null);
+    try {
+      await authManager.api.transferSubscription(sourceOrgId, targetOrgId);
+      setTransfer(null);
+      await refreshAll();
+      toast(`Pro moved to ${targetName}.`);
+    } catch (e) {
+      const message = e instanceof Error ? e.message : String(e);
+      setError(message);
+      toast(`Couldn't move the subscription — ${message}`, "error");
+    }
+  };
+
+  const runCancelNow = async () => {
+    if (!cancelling) return;
+    const { orgId: target, label } = cancelling;
+    setError(null);
+    try {
+      await authManager.api.cancelSubscription(target, "now");
+      setCancelling(null);
+      await refreshAll();
+      toast(`Subscription for ${label} canceled.`);
+    } catch (e) {
+      const message = e instanceof Error ? e.message : String(e);
+      setError(message);
+      toast(`Couldn't cancel the subscription — ${message}`, "error");
+    }
+  };
+
+  if (!billingConfig?.enabled) {
+    return (
+      <div className="muted perm-empty">Billing isn't enabled on this server.</div>
+    );
+  }
+
+  const vaults = myBilling?.vaults ?? [];
+  const orphaned = myBilling?.orphaned ?? [];
+  const freeLimits = myBilling?.freeLimits ?? null;
+
+  /**
+   * The Transfer control for one row. It opens the transfer dialog, where the
+   * eligible destinations are laid out with their seats and plan and the move
+   * is confirmed in the same place. With exactly one eligible vault it is
+   * pre-picked, so the common case is still one click plus a confirm.
+   *
+   * No eligible vault ⇒ a disabled control that says why, rather than a menu
+   * with nothing in it.
+   */
+  const transferControl = (sourceOrgId: string, sourceLabel: string) => {
+    const targets = transferTargets(vaults, sourceOrgId);
+    if (targets.length === 0) {
+      return (
+        <button
+          className="link-btn"
+          disabled
+          title="Nowhere to move it — you need another vault you own that isn't already on Pro."
+        >
+          Transfer
+        </button>
+      );
+    }
+    return (
+      <button
+        type="button"
+        className="link-btn"
+        disabled={busy}
+        aria-label={`Move ${sourceLabel}'s subscription to another vault`}
+        onClick={() => {
+          setError(null);
+          setTransfer({
+            sourceOrgId,
+            sourceLabel,
+            targetOrgId: targets.length === 1 ? targets[0].orgId : null,
+          });
+        }}
+      >
+        Transfer
+      </button>
+    );
+  };
+
+  /**
+   * The transfer dialog body: the eligible destinations as selectable cards.
+   * Each card carries what the reader weighs when choosing — the vault's
+   * seats, and that it is on Free today — instead of a bare name.
+   */
+  const renderTransferDialog = () => {
+    if (!transfer) return null;
+    const targets = transferTargets(vaults, transfer.sourceOrgId);
+    const target = targets.find((t) => t.orgId === transfer.targetOrgId) ?? null;
+    // Owned vaults that are NOT offered, so the list's gaps are explained.
+    const skipped = vaults.filter(
+      (v) => v.orgId !== transfer.sourceOrgId && v.role === "owner" && !targets.includes(v),
+    ).length;
+    return (
+      <ConfirmDialog
+        tone="accent"
+        title={`Move Pro from ${transfer.sourceLabel}`}
+        confirmLabel={target ? `Move Pro to ${target.name}` : "Move subscription"}
+        confirmDisabled={!target}
+        onCancel={() => setTransfer(null)}
+        onConfirm={runTransfer}
+      >
+        <p>
+          Choose the vault that becomes Pro. It keeps the same billing period and
+          price. <strong>{transfer.sourceLabel}</strong> drops to Free — its members
+          and notes stay, but free-plan limits apply to it again.
+        </p>
+        <div className="transfer-targets" role="radiogroup" aria-label="Destination vault">
+          {targets.map((t) => {
+            const used = t.seats.members + t.seats.pendingInvitations;
+            const selected = t.orgId === transfer.targetOrgId;
+            return (
+              <button
+                key={t.orgId}
+                type="button"
+                role="radio"
+                aria-checked={selected}
+                className={`upgrade-plan-card transfer-target${selected ? " selected" : ""}`}
+                onClick={() => setTransfer({ ...transfer, targetOrgId: t.orgId })}
+              >
+                <span className="transfer-target-name">{t.name}</span>
+                <span className="transfer-target-meta">
+                  {used} of {t.seats.limit ?? "∞"} member{used === 1 ? "" : "s"} · Free
+                </span>
+              </button>
+            );
+          })}
+        </div>
+        {skipped > 0 && (
+          <p className="transfer-targets-note">
+            Only vaults you own that aren't already on Pro are listed.
+          </p>
+        )}
+        {error && <div className="auth-error">{error}</div>}
+      </ConfirmDialog>
+    );
+  };
+
+  /** Section 1 — the vault currently open. A plain render helper, NOT a nested
+   *  component: a component declared in here would remount its whole subtree
+   *  on every state change of this page. */
+  const renderVaultCard = () => {
+    if (!isSynced || !orgId) {
+      // A card, not a bare line: this slot holds the Pro/Free card in every
+      // other state, and a naked sentence there left the tab starting on
+      // nothing and the Subscriptions list looking like the whole page.
+      return (
+        <div className="billing-card">
+          <div className="muted">
+            This vault isn't synced, so it has no plan of its own. The vaults on
+            your account are listed below.
+          </div>
+        </div>
+      );
+    }
+    if (!orgBilling) return <div className="muted">Loading…</div>;
+
+    if (orgBilling.plan === "pro") {
+      return (
+        <div className="billing-card plan-pro">
+          <div className="billing-plan-head">
+            <span className="billing-plan-name">Pro</span>
+            <span className={`billing-status ${orgBilling.status}`}>
+              {orgBilling.status === "past_due"
+                ? "Past due"
+                : orgBilling.status === "canceled"
+                  ? "Canceled"
+                  : "Active"}
+            </span>
+          </div>
+          <div className="muted">Everything unlimited on this vault.</div>
+          {orgBilling.currentPeriodEnd && (
+            <div className="menu-row">
+              <span className="menu-row-label">
+                {orgBilling.cancelAtPeriodEnd ? "Access until" : "Renews"}
+              </span>
+              <span>{formatDate(orgBilling.currentPeriodEnd)}</span>
+            </div>
+          )}
+          {orgBilling.cancelAtPeriodEnd && (
+            <div className="limit-nudge">
+              <span>
+                Your subscription is set to cancel at the end of the current period.
+              </span>
+            </div>
+          )}
+          {canManage ? (
+            <AsyncButton
+              className="secondary billing-action"
+              disabled={busy}
+              onClick={() => openPortal(orgId)}
+            >
+              Manage subscription
+            </AsyncButton>
+          ) : (
+            <div className="muted">Ask an owner or admin to manage the subscription.</div>
+          )}
+        </div>
+      );
+    }
+
+    const { members, pendingInvitations, limit } = orgBilling.seats;
+    const used = members + pendingInvitations;
+    return (
+      <div className="billing-card">
+        <div className="billing-plan-head">
+          <span className="billing-plan-name">Free</span>
+        </div>
+        <div className="menu-row">
+          <span className="menu-row-label">Members</span>
+          <span>
+            {used} of {limit ?? "∞"}
+            {limit != null && used >= limit ? " · full" : ""}
+          </span>
+        </div>
+        {pendingInvitations > 0 && (
+          <div className="muted">
+            Includes {pendingInvitations} pending invitation
+            {pendingInvitations === 1 ? "" : "s"}.
+          </div>
+        )}
+
+        <div className="subhead">Upgrade to Pro unlocks</div>
+        <ul className="upgrade-features">
+          <li>Unlimited team members</li>
+          <li>Unlimited notes, devices &amp; AI edits</li>
+          <li>Doesn't count toward your free vaults</li>
+          <li>Priority support</li>
+        </ul>
+
+        {canManage ? (
+          <button className="primary billing-action" onClick={() => setUpgradeOrg(orgId)}>
+            Upgrade to Pro
+          </button>
+        ) : (
+          <div className="muted">Ask an owner or admin to upgrade this vault.</div>
+        )}
+      </div>
+    );
+  };
+
+  /** One row of section 2. */
+  const renderVaultRow = (v: MyBillingVault) => {
+    const seatsUsed = v.seats.members + v.seats.pendingInvitations;
+    // One meta line, not three: renewal + price (already joined by
+    // `subscriptionStatusLine`), then seats, then who pays when that isn't the
+    // reader. A row is a name and a fact line, so the list scans vertically.
+    const meta = [
+      subscriptionStatusLine(v, LINE_FORMAT),
+      `${seatsUsed} of ${v.seats.limit ?? "∞"} member${seatsUsed === 1 ? "" : "s"}`,
+      !v.canManage && v.billingOwner ? `Billed to ${v.billingOwner.name}` : null,
+    ]
+      .filter(Boolean)
+      .join(" · ");
+    return (
+      <li key={v.orgId} className="billing-sub-row">
+        <span className="billing-sub-name">
+          <span className="billing-sub-title">
+            <span className="billing-sub-vault">{v.name}</span>
+            {/* A pill, not muted trailing text: it labels the row the reader
+                arrived from, so it has to survive the name's ellipsis. */}
+            {v.orgId === orgId && <span className="member-role">Current</span>}
+          </span>
+          {meta && <span className="billing-sub-meta">{meta}</span>}
+        </span>
+        {/* Owner is the default for a vault you are billed for, and the row
+            already carries Current + plan pills plus up to three actions — so
+            the role pill only appears when the role is worth saying. */}
+        {v.role !== "owner" && <span className={`member-role ${v.role}`}>{v.role}</span>}
+        <span className={`billing-status ${v.status}`}>{planPillLabel(v)}</span>
+        <span className="vault-row-actions">
+          {v.canManage && v.plan === "free" && (
+            <AsyncButton
+              className="link-btn"
+              disabled={busy}
+              onClick={() => setUpgradeOrg(v.orgId)}
+            >
+              Upgrade
+            </AsyncButton>
+          )}
+          {v.canManage && v.plan === "pro" && (
+            <AsyncButton
+              className="link-btn"
+              disabled={busy}
+              onClick={() => openPortal(v.orgId)}
+            >
+              Manage
+            </AsyncButton>
+          )}
+          {v.canTransfer && transferControl(v.orgId, v.name)}
+        </span>
+      </li>
+    );
+  };
+
+  return (
+    <>
+      {renderVaultCard()}
+
+      {/* ---- 2. Every vault on the account ---- */}
+      <div className="subhead">Subscriptions</div>
+      {vaults.length === 0 ? (
+        <div className="muted perm-empty">
+          {myBilling ? "No vaults on this account yet." : "Loading…"}
+        </div>
+      ) : (
+        <ul className="member-list">{vaults.map(renderVaultRow)}</ul>
+      )}
+
+      {/* ---- 3. Tombstones: paid time that outlived its vault ---- */}
+      {orphaned.length > 0 && (
+        <>
+          <div className="subhead">From deleted vaults</div>
+          <div className="billing-section-note">
+            These subscriptions belonged to vaults that were deleted. They still
+            bill until they end — move one to a vault to use the time you've paid
+            for, or cancel it now.
+          </div>
+          <ul className="member-list">
+            {orphaned.map((o) => {
+              const label = o.orgName ?? "Deleted vault";
+              // Same two-line shape as a live vault row: the name leads, and
+              // when it was deleted is a fact on the meta line, not a
+              // parenthetical that competes with the name for the ellipsis.
+              const meta = [
+                `Deleted ${formatDate(o.deletedAt)}`,
+                subscriptionStatusLine(o, LINE_FORMAT),
+              ]
+                .filter(Boolean)
+                .join(" · ");
+              return (
+                <li key={o.orgId} className="billing-sub-row">
+                  <span className="billing-sub-name">
+                    <span className="billing-sub-title">
+                      <span className="billing-sub-vault">{label}</span>
+                    </span>
+                    <span className="billing-sub-meta">{meta}</span>
+                  </span>
+                  <span className={`billing-status ${o.status}`}>
+                    {o.status === "past_due" ? "Past due" : "Pro"}
+                  </span>
+                  <span className="vault-row-actions">
+                    {transferControl(o.orgId, label)}
+                    <AsyncButton
+                      className="link-btn danger"
+                      disabled={busy}
+                      onClick={() => {
+                        setError(null);
+                        setCancelling({ orgId: o.orgId, label });
+                      }}
+                    >
+                      Cancel now
+                    </AsyncButton>
+                    <AsyncButton
+                      className="link-btn"
+                      disabled={busy}
+                      onClick={() => openPortal(o.orgId)}
+                    >
+                      Manage
+                    </AsyncButton>
+                  </span>
+                </li>
+              );
+            })}
+          </ul>
+        </>
+      )}
+
+      {/* ---- 4. What the free tier allows ---- */}
+      {freeLimits && (
+        <div className="menu-row settings-footer-row">
+          <span className="menu-row-label">Free vaults</span>
+          <span>
+            {freeLimits.freeVaultsUsed} of {freeLimits.vaultsPerUser} used
+          </span>
+        </div>
+      )}
+
+      {error && <div className="auth-error">{error}</div>}
+
+      {upgradeOrg && (
+        <UpgradeDialog orgId={upgradeOrg} onClose={() => setUpgradeOrg(null)} />
+      )}
+
+      {renderTransferDialog()}
+
+      {cancelling && (
+        <ConfirmDialog
+          title={`Cancel the subscription for ${cancelling.label}?`}
+          confirmLabel="Cancel subscription"
+          cancelLabel="Keep it"
+          onCancel={() => setCancelling(null)}
+          onConfirm={runCancelNow}
+        >
+          <p>
+            Billing stops now and the rest of the period is given up. There is no
+            vault left to use it on, so nothing else is lost.
+          </p>
+          <p>
+            If you'd rather keep the time you've paid for, move it to another vault
+            instead.
+          </p>
+          {error && <div className="auth-error">{error}</div>}
+        </ConfirmDialog>
+      )}
+    </>
+  );
+}
+
+/** Compact absolute date for renewal/period-end lines. */
+function formatDate(iso: string): string {
+  const d = new Date(iso);
+  if (Number.isNaN(d.getTime())) return iso;
+  return d.toLocaleDateString(undefined, { year: "numeric", month: "short", day: "numeric" });
+}
+
+/**
+ * How a subscription row writes its date and price. Both formatters already
+ * exist — reused here rather than re-implemented so `lib/billing.ts` can stay
+ * pure and there is exactly one definition of how we print money.
+ */
+const LINE_FORMAT = {
+  date: formatDate,
+  price: (amount: number, currency: string, interval: "month" | "year" | null) =>
+    `${formatPrice({ amount, currency })}${interval ? perLabel(interval) : ""}`,
+};
+
+/**
+ * Inline upgrade nudge shown in the create-vault / invite-member error slot
+ * when the server rejects with a 402 free-plan limit. Styled with --warning-soft
+ * (reserved for upgrade nudges), not the danger palette — this isn't an error.
+ */
+function LimitNudge({
+  kind,
+  limit,
+  onUpgrade,
+}: {
+  kind: LimitKind;
+  limit: number | null;
+  onUpgrade: () => void;
+}) {
+  const freeLimits = useStore((s) => s.billingConfig?.freeLimits);
+  // The server is the authority (the 402 carries `limit`, and billingConfig
+  // reports both caps); these are last-resort defaults for a nudge rendered
+  // before either arrived. Kept separate per kind so the two caps can move
+  // independently (members were 10 for a while; both are 3 since 2026-09-09).
+  const n =
+    limit ??
+    (kind === "member_limit"
+      ? (freeLimits?.membersPerVault ?? 3)
+      : (freeLimits?.vaultsPerUser ?? 3));
+  const message =
+    kind === "member_limit"
+      ? `Free plan limit reached — this vault allows ${n} member${n === 1 ? "" : "s"}.`
+      : // The cap counts FREE vaults only: a Pro vault leaves the count, so
+        // upgrading one of them opens a slot for another free vault.
+        `You're using all ${n} free vault${n === 1 ? "" : "s"}. Upgrade one to Pro — Pro vaults don't count toward that limit — and you can create another.`;
+  return (
+    <div className="limit-nudge">
+      <span>{message}</span>
+      <button className="link-btn" onClick={onUpgrade}>
+        Upgrade →
+      </button>
+    </div>
+  );
+}
+
+/**
+ * Versioning: the vault-wide safety net. Lists the vault's checkpoints (max 5 —
+ * a daily automatic one plus manual ones), and lets an owner OR admin take one
+ * and roll the whole vault back to it. Per-note history lives in the editor's
+ * version panel; this page is for the blast-radius case ("the reorg went wrong,
+ * put everything back").
+ *
+ * Revert used to be owner-only. It is the recovery half of an action admins
+ * could already take (create/delete a checkpoint), so a team whose owner was
+ * away could take checkpoints and not use them.
+ */
+function VersioningTab({ canManage }: { canManage: boolean }) {
+  const checkpoints = useStore((s) => s.checkpoints);
+  const [label, setLabel] = useState("");
+  const [error, setError] = useState<string | null>(null);
+  const [confirming, setConfirming] = useState<VaultCheckpoint | null>(null);
+  const [now, setNow] = useState(() => Date.now());
+
+  useEffect(() => {
+    void useStore.getState().refreshCheckpoints();
+    const id = window.setInterval(() => setNow(Date.now()), 30_000);
+    return () => window.clearInterval(id);
+  }, []);
+
+  const create = async () => {
+    setError(null);
+    try {
+      await useStore.getState().createCheckpoint(label);
+      setLabel("");
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    }
+  };
+
+  const remove = async (id: string) => {
+    setError(null);
+    try {
+      await useStore.getState().deleteCheckpoint(id);
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    }
+  };
+
+  const revert = async (cp: VaultCheckpoint) => {
+    setError(null);
+    try {
+      const result = await useStore.getState().revertVaultToCheckpoint(cp.id);
+      setConfirming(null);
+      toast(
+        `Vault reverted — ${result.docsChanged} notes changed, ` +
+          `${result.docsRestored} restored, ${result.docsDeleted} removed`,
+        "success",
+      );
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    }
+  };
+
+  return (
+    <div className="versioning-tab">
+      <div className="muted">
+        A checkpoint captures every note in this vault — content and folder
+        structure. One is taken automatically each day the vault changes; the
+        vault keeps its 5 most recent. Reverting rolls every member's vault back
+        and broadcasts live.
+      </div>
+      <div className="menu-sep" />
+      {canManage && (
+        <>
+          <div className="subhead">Create checkpoint</div>
+          <div className="row invite-bar">
+            <input
+              type="text"
+              placeholder="Label (optional) — e.g. Before the big reorg"
+              value={label}
+              onChange={(e) => setLabel(e.target.value)}
+            />
+            <AsyncButton className="primary" onClick={create}>
+              Create
+            </AsyncButton>
+          </div>
+          <div className="menu-sep" />
+        </>
+      )}
+      <div className="subhead">Checkpoints</div>
+      {error && <div className="auth-error">{error}</div>}
+      {checkpoints == null ? (
+        <div className="muted perm-empty">Loading…</div>
+      ) : checkpoints.length === 0 ? (
+        <div className="muted perm-empty">
+          No checkpoints yet. One is captured automatically within a day of the
+          vault changing{canManage ? ", or create one above" : ""}.
+        </div>
+      ) : (
+        <ul className="checkpoint-list">
+          {checkpoints.map((cp) => (
+            <li key={cp.id} className="checkpoint-row">
+              <span className="checkpoint-main">
+                <span className="checkpoint-title">
+                  {checkpointTitle(cp.label, cp.kind, cp.createdAt, now)}
+                </span>
+                <span className="checkpoint-sub">
+                  {agoFromIso(cp.createdAt, now)} · {noteCountLabel(cp.noteCount)}
+                  {cp.createdByName ? ` · by ${cp.createdByName}` : ""}
+                </span>
+              </span>
+              {canManage && (
+                <button className="link-btn" onClick={() => setConfirming(cp)}>
+                  Revert
+                </button>
+              )}
+              {canManage && (
+                <AsyncButton className="link-btn danger" onClick={() => remove(cp.id)}>
+                  Delete
+                </AsyncButton>
+              )}
+            </li>
+          ))}
+        </ul>
+      )}
+      {!canManage && (
+        <div className="muted checkpoint-note">
+          Only a vault owner or admin can revert to a checkpoint.
+        </div>
+      )}
+      {confirming && (
+        <div className="modal-backdrop" onClick={() => setConfirming(null)}>
+          <div className="modal" onClick={(e) => e.stopPropagation()}>
+            <div className="modal-header">
+              <h2>Revert the entire vault?</h2>
+            </div>
+            <p className="muted">
+              Every note goes back to{" "}
+              <strong>
+                {checkpointTitle(
+                  confirming.label,
+                  confirming.kind,
+                  confirming.createdAt,
+                  now,
+                )}
+              </strong>{" "}
+              ({agoFromIso(confirming.createdAt, now)}) — content, names and
+              folders — for every member, live. Notes created since then are
+              moved to trash. Attachments are not reverted. A checkpoint of the
+              current state is taken first, so this can be undone.
+            </p>
+            <div className="banner-actions">
+              <button className="secondary" onClick={() => setConfirming(null)}>
+                Cancel
+              </button>
+              <AsyncButton
+                className="primary danger"
+                spinnerTone="on-accent"
+                onClick={() => revert(confirming)}
+              >
+                Revert vault
+              </AsyncButton>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
+
+/**
+ * MCP: expose this vault to AI clients over the Model Context Protocol.
+ * The MCP endpoint is part of the same server; a client authenticates with a
+ * token minted here and then gets the SAME CRUD access to notes/folders that
+ * the signed-in user has (owners/admins see everything; members see what's
+ * shared with them). This is where you grab the URL + a token.
+ */
+function McpTab() {
+  const session = useStore((s) => s.session);
+  const serverUrl = useStore((s) => s.serverUrl);
+
+  const mcpUrl = `${serverUrl.replace(/\/+$/, "")}/api/mcp`;
+  const hasVault = !!session?.activeOrganizationId;
+
+  const [tokens, setTokens] = useState<McpTokenRow[]>([]);
+  const [tools, setTools] = useState<McpToolInfo[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [name, setName] = useState("");
+  const [busy, setBusy] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+  const [justCreated, setJustCreated] = useState<{ name: string; token: string } | null>(
+    null,
+  );
+  const [copied, setCopied] = useState<string | null>(null);
+  const [expanded, setExpanded] = useState<string | null>(null);
+  // Bumps every 20s so "connected" dots + relative times stay live while open.
+  const [, setTick] = useState(0);
+
+  useEffect(() => {
+    if (!hasVault) {
+      setLoading(false);
+      return;
+    }
+    let cancelled = false;
+    const load = () =>
+      authManager.api
+        .listMcpConnections()
+        .then(({ tokens, tools }) => {
+          if (cancelled) return;
+          setTokens(tokens);
+          setTools(tools);
+        })
+        .catch(() => {})
+        .finally(() => {
+          if (!cancelled) setLoading(false);
+        });
+    void load();
+    // Poll so a connection that goes active/idle while the panel is open shows it.
+    const poll = window.setInterval(() => void load(), 20_000);
+    const tickle = window.setInterval(() => setTick((n) => n + 1), 20_000);
+    return () => {
+      cancelled = true;
+      window.clearInterval(poll);
+      window.clearInterval(tickle);
+    };
+  }, [hasVault]);
+
+  const copy = async (text: string, tag: string) => {
+    try {
+      await navigator.clipboard.writeText(text);
+      setCopied(tag);
+      window.setTimeout(() => setCopied((c) => (c === tag ? null : c)), 1500);
+    } catch {
+      /* clipboard unavailable */
+    }
+  };
+
+  const create = async () => {
+    setBusy(true);
+    setError(null);
+    try {
+      const created = await authManager.api.createMcpToken(name.trim() || "MCP token");
+      setJustCreated({ name: created.name, token: created.token });
+      const { token: _t, ...row } = created;
+      setTokens((prev) => [row, ...prev]);
+      setName("");
+    } catch (e) {
+      const message = e instanceof Error ? e.message : String(e);
+      setError(message);
+      // Otherwise the button just returns to idle and the user clicks again,
+      // minting duplicate tokens on a server that is actually failing (#85).
+      toast(`Couldn't create the MCP token — ${message}`, "error");
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const revoke = async (id: string) => {
+    setBusy(true);
+    setError(null);
+    try {
+      await authManager.api.revokeMcpToken(id);
+      setTokens((prev) => prev.filter((t) => t.id !== id));
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  if (!hasVault) {
+    return (
+      <div className="muted perm-empty">
+        MCP needs an active vault — create or switch to one first.
+      </div>
+    );
+  }
+
+  const snippet = justCreated
+    ? `claude mcp add --transport http context ${mcpUrl} \\\n  --header "Authorization: Bearer ${justCreated.token}"`
+    : "";
+
+  return (
+    <>
+      <div className="muted">
+        Connect any MCP-compatible AI client to this vault. It gets the same
+        access you do — read, search, create, edit and delete notes and folders.
+      </div>
+
+      <div className="subhead">Endpoint URL</div>
+      <div className="join-code-row">
+        <code className="vault-root-path" title={mcpUrl}>
+          {mcpUrl}
+        </code>
+        <button className="link-btn" onClick={() => void copy(mcpUrl, "url")}>
+          {copied === "url" ? "Copied ✓" : "Copy"}
+        </button>
+      </div>
+
+      <div className="menu-sep" />
+      <div className="subhead">Access tokens</div>
+      <div className="muted">
+        A token authenticates the client and scopes it to you in this vault.
+        Add it as an <code>Authorization: Bearer</code> header. Revoke any time.
+      </div>
+
+      <div className="row invite-bar">
+        <input
+          placeholder="Token name, e.g. Claude Desktop"
+          value={name}
+          onChange={(e) => setName(e.target.value)}
+          onKeyDown={(e) => {
+            if (e.key === "Enter") void create();
+          }}
+        />
+        <button className="primary" disabled={busy} onClick={() => void create()}>
+          Create token
+        </button>
+      </div>
+
+      {error && <div className="auth-error">{error}</div>}
+
+      {justCreated && (
+        <div className="mcp-new-token">
+          <div className="subhead">Copy your token now — it won't be shown again</div>
+          <div className="join-code-row">
+            <code className="join-code mcp-token-value" title={justCreated.token}>
+              {justCreated.token}
+            </code>
+            <button
+              className="link-btn"
+              onClick={() => void copy(justCreated.token, "token")}
+            >
+              {copied === "token" ? "Copied ✓" : "Copy"}
+            </button>
+          </div>
+          <div className="muted">Example — add it to Claude Code:</div>
+          <div className="join-code-row">
+            <code className="mcp-snippet">{snippet}</code>
+            <button className="link-btn" onClick={() => void copy(snippet, "snippet")}>
+              {copied === "snippet" ? "Copied ✓" : "Copy"}
+            </button>
+          </div>
+          <button className="link-btn" onClick={() => setJustCreated(null)}>
+            Done
+          </button>
+        </div>
+      )}
+
+      <div className="menu-sep" />
+      <div className="subhead">Connections</div>
+      <div className="muted">
+        Every token is a connection into this vault. Each reaches the same{" "}
+        {tools.length || ""} tools, gated by your access — expand one to see them,
+        how active it is, and how much it's been used.
+      </div>
+
+      {loading ? (
+        <div className="muted">Loading…</div>
+      ) : tokens.length === 0 ? (
+        <div className="muted">No connections yet.</div>
+      ) : (
+        <ul className="mcp-conn-list">
+          {tokens.map((t) => {
+            const live = isConnected(t.lastUsedAt);
+            const open = expanded === t.id;
+            return (
+              <li key={t.id} className={`mcp-conn${open ? " open" : ""}`}>
+                <div className="mcp-conn-head">
+                  <span
+                    className={`mcp-dot ${live ? "on" : "off"}`}
+                    title={live ? "Connected" : "Disconnected"}
+                    aria-hidden="true"
+                  />
+                  <div className="mcp-conn-main">
+                    <div className="mcp-conn-title">
+                      {t.name}
+                      <span className={`mcp-status ${live ? "on" : "off"}`}>
+                        {live ? "Connected" : "Disconnected"}
+                      </span>
+                    </div>
+                    <div className="mcp-conn-sub muted">
+                      {clientLabel(t.lastClient)}
+                      {" · "}
+                      {t.tokenPrefix}
+                      {" · "}
+                      {t.useCount} {t.useCount === 1 ? "call" : "calls"}
+                      {" · "}
+                      {t.lastUsedAt ? `last active ${relTime(t.lastUsedAt)}` : "never used"}
+                    </div>
+                  </div>
+                  <button
+                    className="link-btn"
+                    onClick={() => setExpanded((e) => (e === t.id ? null : t.id))}
+                  >
+                    {open ? "Hide tools" : `Tools · ${tools.length}`}
+                  </button>
+                  <AsyncButton
+                    className="link-btn danger"
+                    disabled={busy}
+                    onClick={() => revoke(t.id)}
+                  >
+                    Revoke
+                  </AsyncButton>
+                </div>
+                {open && (
+                  <ul className="mcp-tool-list">
+                    {tools.map((tool) => (
+                      <li key={tool.name} title={tool.description}>
+                        <span className={`mcp-tool-badge ${tool.access}`}>
+                          {tool.access === "read"
+                            ? "read"
+                            : tool.access === "destructive"
+                              ? "delete"
+                              : "write"}
+                        </span>
+                        <code>{tool.name}</code>
+                      </li>
+                    ))}
+                  </ul>
+                )}
+              </li>
+            );
+          })}
+        </ul>
+      )}
+    </>
+  );
+}
+
+/** A connection is "active" when it made a request in the last few minutes
+ *  (MCP here is stateless HTTP — there's no socket to watch, so recency is it). */
+const CONNECTED_WINDOW_MS = 3 * 60 * 1000;
+function isConnected(lastUsedAt: string | null): boolean {
+  if (!lastUsedAt) return false;
+  return Date.now() - new Date(lastUsedAt).getTime() < CONNECTED_WINDOW_MS;
+}
+
+/** Compact relative time: "just now", "5m ago", "3h ago", "2d ago", else a date. */
+function relTime(iso: string): string {
+  const diff = Date.now() - new Date(iso).getTime();
+  if (diff < 60_000) return "just now";
+  const m = Math.floor(diff / 60_000);
+  if (m < 60) return `${m}m ago`;
+  const h = Math.floor(m / 60);
+  if (h < 24) return `${h}h ago`;
+  const d = Math.floor(h / 24);
+  if (d < 7) return `${d}d ago`;
+  return new Date(iso).toLocaleDateString();
+}
+
+/** Best-effort human name for a client from its User-Agent. */
+function clientLabel(ua: string | null): string {
+  if (!ua) return "Unknown client";
+  const s = ua.toLowerCase();
+  if (s.includes("claude-code") || s.includes("claude code")) return "Claude Code";
+  if (s.includes("claude")) return "Claude";
+  if (s.includes("cursor")) return "Cursor";
+  if (s.includes("node")) return "Node client";
+  // Fall back to the leading token of the UA (e.g. "MyApp/1.2" → "MyApp").
+  return ua.split(/[\s/]/)[0].slice(0, 40) || "Unknown client";
+}
+
+/**
+ * Appearance: theme plus the vault's folder/note colors. Colors are assigned
+ * from each item's ⋯ menu in the sidebar; this tab reviews and clears them.
+ */
+/**
+ * Updates tab — shows the running version and lets the user check for and
+ * install a newer release on demand. The launch-time check populates the same
+ * shared updater state, so if an update was already found this reflects it.
+ */
+function UpdatesTab() {
+  const update = useUpdateState();
+  const [version, setVersion] = useState<string | null>(null);
+
+  useEffect(() => {
+    void currentVersion().then(setVersion);
+  }, []);
+
+  const busy = update.phase === "checking" ||
+    update.phase === "downloading" ||
+    update.phase === "installing";
+
+  // A single status line that stays mounted across phases so the card never
+  // reflows (and the button never jumps) as the check progresses. The button
+  // keeps one fixed label + width; the spinner and this line carry the state.
+  let statusText: string | null = null;
+  let statusError = false;
+  switch (update.phase) {
+    case "checking":
+      statusText = "Checking for updates…";
+      break;
+    case "uptodate":
+      statusText = "You're on the latest version.";
+      break;
+    case "available":
+      statusText = "An update is available.";
+      break;
+    case "downloading":
+      statusText = update.total > 0
+        ? `Downloading ${update.version} — ${Math.round((update.downloaded / update.total) * 100)}%`
+        : `Downloading ${update.version}…`;
+      break;
+    case "installing":
+      statusText = `Installing ${update.version} — the app will restart…`;
+      break;
+    case "error":
+      statusText = `Couldn't check for updates: ${update.message}`;
+      statusError = true;
+      break;
+  }
+
+  return (
+    <div className="updates-tab">
+      <div className="menu-row">
+        <span className="menu-row-label">Current version</span>
+        <span className="mono">{version ?? "…"}</span>
+      </div>
+
+      <div className="update-actions">
+        <button
+          className="primary sm update-check-btn"
+          disabled={busy}
+          aria-busy={busy}
+          onClick={() => void checkForUpdate()}
+        >
+          {busy && <span className="btn-spinner" aria-hidden="true" />}
+          <span>Check for updates</span>
+        </button>
+
+        <span
+          className={`update-status${statusError ? " error" : ""}`}
+          role="status"
+          aria-live="polite"
+        >
+          {statusText}
+        </span>
+      </div>
+
+      {update.phase === "available" && (
+        <div className="update-detail">
+          <div className="subhead">Version {update.version} available</div>
+          {update.notes && <div className="muted release-notes">{update.notes}</div>}
+          <AsyncButton className="primary sm" onClick={() => installUpdate()}>
+            Install &amp; Restart
+          </AsyncButton>
+        </div>
+      )}
+    </div>
+  );
+}
+
+const APPEARANCE_ICON = {
+  folder: (
+    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round">
+      <path d="M3 7a2 2 0 0 1 2-2h4l2 2h6a2 2 0 0 1 2 2v8a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" />
+    </svg>
+  ),
+  note: (
+    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round">
+      <path d="M7 3h7l5 5v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V4a1 1 0 0 1 1-1z" />
+      <path d="M14 3v5h5" />
+    </svg>
+  ),
+};
+
+function importSummaryText(s: ipc.ImportSummary): string {
+  const parts = [`Imported ${s.files} file${s.files === 1 ? "" : "s"}`];
+  if (s.skipped > 0) parts.push(`${s.skipped} skipped`);
+  return parts.join(" · ") + ".";
+}
+
+/**
+ * Import / Export — vault-level data operations on the open local vault. Imports
+ * land at the vault root; exports copy out to a chosen folder. The same commands
+ * back the sidebar ⋮ menu and drag-and-drop, so behavior is identical everywhere.
+ */
+function ImportExportTab() {
+  const [busy, setBusy] = useState<null | "files" | "folder" | "export">(null);
+  const [msg, setMsg] = useState<string | null>(null);
+  const [error, setError] = useState<string | null>(null);
+
+  async function refresh() {
+    await useStore.getState().refreshTree();
+    await useStore.getState().refreshTitles();
+  }
+
+  async function run(
+    kind: "files" | "folder" | "export",
+    fn: () => Promise<string | null>,
+  ) {
+    setBusy(kind);
+    setError(null);
+    setMsg(null);
+    try {
+      const result = await fn();
+      if (result) setMsg(result);
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setBusy(null);
+    }
+  }
+
+  const importFiles = () =>
+    run("files", async () => {
+      // Captured before the native picker — a vault switch while it is open must
+      // not redirect the import into the vault the user landed in.
+      const epoch = useStore.getState().vault?.epoch;
+      const sources = await ipc.pickFiles();
+      if (!sources || sources.length === 0) return null;
+      const summary = await ipc.importPaths("", sources, epoch);
+      await refresh();
+      return importSummaryText(summary);
+    });
+
+  const importFolder = () =>
+    run("folder", async () => {
+      const epoch = useStore.getState().vault?.epoch; // before the dialog
+      const src = await ipc.pickFolder();
+      if (!src) return null;
+      const summary = await ipc.importPaths("", [src], epoch);
+      await refresh();
+      return importSummaryText(summary);
+    });
+
+  const exportVault = () =>
+    run("export", async () => {
+      const epoch = useStore.getState().vault?.epoch; // before the dialog
+      const dest = await ipc.pickFolder();
+      if (!dest) return null;
+      await ipc.exportPath("", dest, epoch);
+      return "Exported the vault.";
+    });
+
+  return (
+    <div className="io-tab">
+      <section className="io-section">
+        <h3 className="io-heading">Import</h3>
+        <p className="io-desc">
+          Bring existing files and folders into this vault — any format. Markdown and text
+          become notes; everything else is kept as-is, with its folder structure. Existing
+          names are never overwritten.
+        </p>
+        <div className="io-actions">
+          <button className="primary" disabled={busy !== null} onClick={() => void importFiles()}>
+            {busy === "files" ? "Importing…" : "Import files…"}
+          </button>
+          <button className="primary" disabled={busy !== null} onClick={() => void importFolder()}>
+            {busy === "folder" ? "Importing…" : "Import folder…"}
+          </button>
+        </div>
+        <p className="io-hint">
+          You can also right-click any folder in the sidebar, or drag files straight onto it.
+        </p>
+      </section>
+
+      <section className="io-section">
+        <h3 className="io-heading">Export</h3>
+        <p className="io-desc">
+          Save a copy of this whole vault to a folder on your computer. The hidden{" "}
+          <code>.context</code> index is skipped.
+        </p>
+        <div className="io-actions">
+          <button className="primary" disabled={busy !== null} onClick={() => void exportVault()}>
+            {busy === "export" ? "Exporting…" : "Export entire vault…"}
+          </button>
+        </div>
+      </section>
+
+      {error ? (
+        <div className="auth-error">{error}</div>
+      ) : (
+        msg && <div className="io-result">{msg}</div>
+      )}
+    </div>
+  );
+}
+
+function AppearanceTab() {
+  const itemColors = useStore((s) => s.itemColors);
+  const tree = useStore((s) => s.tree);
+
+  // Flatten the vault into indented rows, same order as the sidebar.
+  const items = useMemo(() => {
+    const out: Array<{ path: string; name: string; depth: number; isDir: boolean }> = [];
+    const walk = (n: ipc.TreeNode, depth: number) => {
+      out.push({
+        path: n.path,
+        name: n.isDir ? n.name : n.name.replace(/\.(md|html?)$/i, ""),
+        depth,
+        isDir: n.isDir,
+      });
+      n.children?.forEach((c) => walk(c, depth + 1));
+    };
+    tree?.children?.forEach((c) => walk(c, 0));
+    return out;
+  }, [tree]);
+
+  const coloredCount = items.filter((i) => itemColors[i.path]).length;
+
+  return (
+    <>
+      <div className="menu-row">
+        <span className="menu-row-label">Theme</span>
+        <ThemeToggle />
+      </div>
+
+      <div className="subhead">Folder &amp; note colors</div>
+      <div className="muted">
+        Color-code your sidebar: click a swatch to tint that folder or note. Colors are saved
+        with this device's vault settings.
+      </div>
+
+      {items.length === 0 ? (
+        <div className="muted perm-empty">Open a vault to color its folders and notes.</div>
+      ) : (
+        <>
+          <ul className="appearance-list">
+            {items.map((item) => {
+              const active = itemColors[item.path];
+              return (
+                <li
+                  key={item.path}
+                  className="appearance-row"
+                  style={{ paddingLeft: `${12 + item.depth * 16}px` }}
+                >
+                  <span
+                    className="appearance-glyph"
+                    style={{ color: itemColorValue(active) }}
+                    aria-hidden="true"
+                  >
+                    {item.isDir ? APPEARANCE_ICON.folder : APPEARANCE_ICON.note}
+                  </span>
+                  <span className="appearance-name" title={item.path}>
+                    {item.name}
+                  </span>
+                  <span className="appearance-swatches" role="radiogroup" aria-label={`Color for ${item.name}`}>
+                    <button
+                      type="button"
+                      className={`swatch clear${!active ? " on" : ""}`}
+                      title="Default"
+                      aria-label="Default color"
+                      onClick={() => useStore.getState().setItemColor(item.path, null)}
+                    />
+                    {ITEM_COLORS.map((c) => (
+                      <button
+                        key={c.id}
+                        type="button"
+                        className={`swatch${active === c.id ? " on" : ""}`}
+                        style={{ backgroundColor: c.value }}
+                        title={c.label}
+                        aria-label={c.label}
+                        onClick={() => useStore.getState().setItemColor(item.path, c.id)}
+                      />
+                    ))}
+                  </span>
+                </li>
+              );
+            })}
+          </ul>
+          {coloredCount > 0 && (
+            <button
+              className="link-btn"
+              onClick={() => {
+                const { itemColors: colors, setItemColor } = useStore.getState();
+                Object.keys(colors).forEach((path) => setItemColor(path, null));
+              }}
+            >
+              Clear all colors ({coloredCount})
+            </button>
+          )}
+        </>
+      )}
+    </>
+  );
+}
